--- src/main/java/com/shootdoori/match/entity/user/DeviceType.java
@@ -1,29 +1,0 @@
-package com.shootdoori.match.entity.user;
-
-import org.springframework.util.StringUtils;
-
-public enum DeviceType {
-    ANDROID,
-    IOS,
-    WEB,
-    UNKNOWN;
-
-    public static DeviceType fromUserAgent(String userAgent) {
-        if (!StringUtils.hasText(userAgent)) {
-            return UNKNOWN;
-        }
-
-        String lower = userAgent.toLowerCase();
-        if (lower.contains("android")) {
-            return ANDROID;
-        }
-        if (lower.contains("iphone") || lower.contains("ipad")) {
-            return IOS;
-        }
-        if (lower.contains("windows") || lower.contains("mac") || lower.contains("linux")) {
-            return WEB;
-        }
-
-        return UNKNOWN;
-    }
-}
