--- src/main/java/com/shootdoori/match/entity/user/GeneratedToken.java
@@ -1,10 +1,0 @@
-package com.shootdoori.match.entity.user;
-
-import com.shootdoori.match.entity.user.dto.AuthToken;
-
-public record GeneratedToken(
-    AuthToken authToken,
-    RefreshToken refreshToken
-) {
-
-}
