--- src/main/java/com/shootdoori/match/entity/user/dto/AuthToken.java
@@ -1,9 +1,0 @@
-package com.shootdoori.match.entity.user.dto;
-
-public record AuthToken(
-    String accessToken,
-    String refreshToken,
-    long accessTokenExpiresIn,
-    long refreshTokenExpiresIn
-) {
-}
