--- src/main/java/com/shootdoori/match/entity/user/dto/AuthTokenResponse.java
@@ -1,9 +1,0 @@
-package com.shootdoori.match.entity.user.dto;
-
-public record AuthTokenResponse(
-        String accessToken,
-        String refreshToken,
-        long accessTokenExpiresIn,
-        long refreshTokenExpiresIn
-) {
-}
