--- src/main/java/com/shootdoori/match/entity/user/dto/ProfileRequest.java
@@ -1,4 +1,0 @@
-package com.shootdoori.match.entity.user.dto;
-
-public record ProfileRequest(Long id) {
-}
