--- src/main/java/com/shootdoori/match/entity/user/repository/RefreshTokenRepository.java
@@ -1,8 +1,0 @@
-package com.shootdoori.match.entity.user.repository;
-
-import com.shootdoori.match.entity.user.RefreshToken;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RefreshTokenRepository extends JpaRepository<RefreshToken, String> {
-    
-}
