+++ src/main/java/com/shootdoori/match/user/dto/AuthToken.java
@@ -1,0 +1,9 @@
+package com.shootdoori.match.user.dto;
+
+public record AuthToken(
+    String accessToken,
+    String refreshToken,
+    long accessTokenExpiresIn,
+    long refreshTokenExpiresIn
+) {
+}
