+++ src/main/java/com/shootdoori/match/user/dto/AuthTokenResponse.java
@@ -1,0 +1,9 @@
+package com.shootdoori.match.user.dto;
+
+public record AuthTokenResponse(
+        String accessToken,
+        String refreshToken,
+        long accessTokenExpiresIn,
+        long refreshTokenExpiresIn
+) {
+}
