+++ src/main/java/com/shootdoori/match/user/dto/ProfileRequest.java
@@ -1,0 +1,4 @@
+package com.shootdoori.match.user.dto;
+
+public record ProfileRequest(Long id) {
+}
