+++ src/main/java/com/shootdoori/match/user/mapper/UserMapper.java
@@ -1,0 +1,27 @@
+package com.shootdoori.match.user.mapper;
+
+import com.shootdoori.match.user.domain.User;
+import com.shootdoori.match.user.dto.ProfileResponse;
+import org.springframework.stereotype.Component;
+
+@Component
+public class UserMapper {
+    public ProfileResponse toProfileResponse(User user) {
+        if (user == null) {
+            return null;
+        }
+
+        return new ProfileResponse(
+            user.getName().value(),
+            user.getSkillLevel().name(),
+            user.getEmail().value(),
+            user.getKakaoTalkId().value(),
+            user.getPosition().name(),
+            user.getUniversity().getUniversityName(),
+            user.getDepartment().value(),
+            user.getStudentYear().value(),
+            user.getBio().value(),
+            user.getTimeStamp().getCreatedAt()
+        );
+    }
+}
