+++ src/main/java/com/shootdoori/match/user/util/GeneratedToken.java
@@ -1,0 +1,11 @@
+package com.shootdoori.match.user.util;
+
+import com.shootdoori.match.user.domain.RefreshToken;
+import com.shootdoori.match.user.dto.AuthToken;
+
+public record GeneratedToken(
+    AuthToken authToken,
+    RefreshToken refreshToken
+) {
+
+}
